Update objectLiteralShorthandRule.ts
The error message `Expected property shorthand in object literal ('{injector}').` confused me because it appeared within an object literal like: `const x = { <lines of properties with initializers>, injector: injector, <more lines of other properties>}`.  As I was unfamiliar with the combination of object literals and short-hand properties, I took the authoritative voice of the lint message literally and attempted  `const x = { <lines of properties with initializers>, {injector}, <more lines of other properties>}`, which of course is silly in retrospect, but that was exactly the suggestion of the error message.

By adding context indicators around the token novice users learn from lint that the syntax for shorthand properties is simple.